OperationIdType redesign for ResetHandler
